refactoring error boundary
